Stylize faculty section
This commit  stylizes the faculty section of the web page. This commit reduces the margins of the header to reduce gaps in the page, removes the background color, makes the images circles, adds yellow border to those circles, and edits the text for the faculty's name and office. Adds comments for examples of class, id , and nesting.